GO-1838: Fix type for column. Restore previous version

+++ persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/Label.java
@@ -13,7 +13,7 @@ public class Label implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Integer i;
+	private byte i;
 
 	@Column(name="name",columnDefinition="char(30)")
 	private String name;
@@ -21,11 +21,11 @@ public class Label implements Serializable {
     public Label() {
     }
 
-	public Integer getI() {
+	public byte getI() {
 		return this.i;
 	}
 
-	public void setI(Integer i) {
+	public void setI(byte i) {
 		this.i = i;
 	}
