+++ persistence/src/main/java/mobi/nowtechnologies/server/persistence/repository/LabelRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 /**
  * @author Titov Mykhaylo (titov) on 25.06.2014.
  */
-public interface LabelRepository extends JpaRepository<Label, Integer> {
+public interface LabelRepository extends JpaRepository<Label, Byte> {
 
     @Query(value = "select a from Label a where a.name = ?1")
     Label findByName(String name);
